update socials url logic
